[#971] Extract Backup email page to a new component
with @tuliocasagrande
