+++ web-ui/app/locales/pt_BR/translation.json
@@ -81,14 +81,16 @@
     },
     "backup-account": {
       "page-title": "Pixelated E-mail de Recuperação",
-      "image-description": "Esqueci minha senha!",
-      "title": "E se você esquecer sua senha?",
-      "paragraph1": "Informe outro e-mail que você usa regularmente. Esse será o seu e-mail de recuperação.",
-      "paragraph2": "Instruções para recuperar sua senha serão enviadas para esse e-mail, guarde com carinho.",
-      "input-label": "Digite seu e-mail de recuperação",
-      "button": "Adicionar e-mail",
-      "error": {
-        "invalid-email": "Por favor informe um e-mail válido"
+      "backup-email": {
+        "image-description": "Esqueci minha senha!",
+        "title": "E se você esquecer sua senha?",
+        "paragraph1": "Informe outro e-mail que você usa regularmente. Esse será o seu e-mail de recuperação.",
+        "paragraph2": "Instruções para recuperar sua senha serão enviadas para esse e-mail, guarde com carinho.",
+        "input-label": "Digite seu e-mail de recuperação",
+        "button": "Adicionar e-mail",
+        "error": {
+          "invalid-email": "Por favor informe um e-mail válido"
+        }
       }
     },
     "back-to-inbox": "Voltar",
