+++ web-ui/package.json
@@ -8,6 +8,7 @@
     "babel": "^6.5.2",
     "babel-cli": "^6.22.2",
     "babel-core": "^6.21.0",
+    "babel-eslint": "^7.1.1",
     "babel-loader": "^6.2.10",
     "babel-plugin-istanbul": "^3.1.2",
     "babel-preset-es2015": "^6.18.0",
@@ -18,7 +19,7 @@
     "css-loader": "^0.26.1",
     "dompurify": "^0.8.4",
     "enzyme": "^2.7.1",
-    "eslint": "^3.15.0",
+    "eslint": "^3.17.1",
     "eslint-config-airbnb": "^14.1.0",
     "eslint-plugin-import": "^2.2.0",
     "eslint-plugin-jsx-a11y": "^4.0.0",
